Reorganize headers. Correct get_default_init_options functions
