+++ rmw/include/rmw/init_options.h
@@ -25,8 +25,41 @@ extern "C"
 #include "rcutils/allocator.h"
 #include "rmw/macros.h"
 #include "rmw/ret_types.h"
-#include "rmw/types.h"
+#include "rmw/security_options.h"
 #include "rmw/visibility_control.h"
+
+/// Implementation defined options structure used during rmw_init().
+/**
+ * This should be defined by the rmw implementation.
+ */
+typedef struct rmw_init_options_impl_t rmw_init_options_impl_t;
+
+#define RMW_INIT_OPTIONS_DEFAULT_DOMAIN_ID SIZE_MAX
+
+/// Options structure used during rmw_init().
+typedef struct RMW_PUBLIC_TYPE rmw_init_options_t
+{
+  /// Locally (process local) unique ID that represents this init/shutdown cycle.
+  /**
+   * This should be set by the caller of `rmw_init()` to a number that is
+   * unique within this process.
+   * It is designed to be used with `rcl_init()` and `rcl_get_instance_id()`.
+   */
+  uint64_t instance_id;
+  /// Implementation identifier, used to ensure two different implementations are not being mixed.
+  const char * implementation_identifier;
+  /// ROS domain id
+  size_t domain_id;
+  /// Security options
+  rmw_security_options_t security_options;
+
+  // TODO(wjwwood): replace with rmw_allocator_t when that refactor happens
+  /// Allocator used during internal allocation of init options, if needed.
+  rcutils_allocator_t allocator;
+  /// Implementation defined init options.
+  /** May be NULL if there are no implementation defined options. */
+  rmw_init_options_impl_t * impl;
+} rmw_init_options_t;
 
 /// Return a zero initialized init options structure.
 RMW_PUBLIC
