+++ rmw/include/rmw/security_options.h
@@ -20,7 +20,19 @@ extern "C"
 {
 #endif
 
-#include "rmw/types.h"
+#include "rmw/visibility_control.h"
+
+enum RMW_PUBLIC_TYPE rmw_security_enforcement_policy_t
+{
+  RMW_SECURITY_ENFORCEMENT_PERMISSIVE,
+  RMW_SECURITY_ENFORCEMENT_ENFORCE,
+};
+
+typedef struct RMW_PUBLIC_TYPE rmw_security_options_t
+{
+  enum rmw_security_enforcement_policy_t enforce_security;
+  const char * security_root_path;
+} rmw_security_options_t;
 
 RMW_PUBLIC
 rmw_security_options_t
