+++ rmw/include/rmw/types.h
@@ -27,32 +27,17 @@ extern "C"
 // map rcutils specific log levels to rmw speicfic type
 #include <rcutils/logging.h>
 
+#include "rmw/init.h"
+#include "rmw/init_options.h"
 #include "rmw/loaned_message_sequence.h"
 #include "rmw/ret_types.h"
+#include "rmw/security_options.h"
 #include "rmw/serialized_message.h"
 #include "rmw/visibility_control.h"
 
 // 24 bytes is the most memory needed to represent the GID by any current
 // implementation. It may need to be increased in the future.
 #define RMW_GID_STORAGE_SIZE 24
-
-/// Implementation defined context structure returned by rmw_init().
-/**
- * This should be defined by the rmw implementation.
- */
-typedef struct rmw_context_impl_t rmw_context_impl_t;
-
-/// Initialization context structure which is used to store init specific information.
-typedef struct RMW_PUBLIC_TYPE rmw_context_t
-{
-  /// Locally (process local) unique ID that represents this init/shutdown cycle.
-  uint64_t instance_id;
-  /// Implementation identifier, used to ensure two different implementations are not being mixed.
-  const char * implementation_identifier;
-  /// Implementation defined context information.
-  /** May be NULL if there is no implementation defined context information. */
-  rmw_context_impl_t * impl;
-} rmw_context_t;
 
 typedef struct RMW_PUBLIC_TYPE rmw_node_t
 {
@@ -269,52 +254,6 @@ typedef struct RMW_PUBLIC_TYPE rmw_time_t
   uint64_t sec;
   uint64_t nsec;
 } rmw_time_t;
-
-enum RMW_PUBLIC_TYPE rmw_security_enforcement_policy_t
-{
-  RMW_SECURITY_ENFORCEMENT_PERMISSIVE,
-  RMW_SECURITY_ENFORCEMENT_ENFORCE,
-};
-
-typedef struct RMW_PUBLIC_TYPE rmw_security_options_t
-{
-  enum rmw_security_enforcement_policy_t enforce_security;
-  const char * security_root_path;
-} rmw_security_options_t;
-
-/// Constant which indicates that the default domain id should be used.
-#define RCL_NODE_OPTIONS_DEFAULT_DOMAIN_ID SIZE_MAX
-
-/// Implementation defined options structure used during rmw_init().
-/**
- * This should be defined by the rmw implementation.
- */
-typedef struct rmw_init_options_impl_t rmw_init_options_impl_t;
-
-/// Options structure used during rmw_init().
-typedef struct RMW_PUBLIC_TYPE rmw_init_options_t
-{
-  /// Locally (process local) unique ID that represents this init/shutdown cycle.
-  /**
-   * This should be set by the caller of `rmw_init()` to a number that is
-   * unique within this process.
-   * It is designed to be used with `rcl_init()` and `rcl_get_instance_id()`.
-   */
-  uint64_t instance_id;
-  /// Implementation identifier, used to ensure two different implementations are not being mixed.
-  const char * implementation_identifier;
-  /// ROS domain id
-  size_t domain_id;
-  /// Security options
-  rmw_security_options_t security_options;
-
-  // TODO(wjwwood): replace with rmw_allocator_t when that refactor happens
-  /// Allocator used during internal allocation of init options, if needed.
-  rcutils_allocator_t allocator;
-  /// Implementation defined init options.
-  /** May be NULL if there are no implementation defined options. */
-  rmw_init_options_impl_t * impl;
-} rmw_init_options_t;
 
 enum RMW_PUBLIC_TYPE rmw_qos_reliability_policy_t
 {
