+++ rmw/src/security_options.c
@@ -12,6 +12,8 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
+#include <stddef.h>
+
 #include "rmw/security_options.h"
 
 rmw_security_options_t
